+++ samples/Arlecchino.Commander/Model/MenuSection.cs
@@ -1,5 +1,0 @@
-using System.Collections.Generic;
-
-namespace Arlecchino.Commander.Model;
-
-public sealed record MenuSection(string Title, IReadOnlyList<string> Items);