+++ samples/Arlecchino.Commander/Model/Sorting.cs
@@ -1,8 +1,0 @@
-namespace Arlecchino.Commander.Model;
-
-public enum Sorting
-{
-    Name,
-    Size,
-    Modified,
-}